requirements.txt
Django==3.1.2
psycopg2-binary==2.8.6
django-cors-headers==3.5.0
django-storages==1.10.1
djangorestframework==3.12.1
django-environ==0.4.5
djangorestframework-camel-case==1.2.0
django-anymail==8.1
gunicorn==20.0.4
uvicorn[standard]==0.12.2
WeasyPrint==51
zygoat-django==0.2.0
